requirements.txt
# Google ADK - Agent Development Kit
google-adk>=1.0.0

# MCP - Model Context Protocol
mcp[cli]>=1.0.0

# FastAPI & Server
fastapi>=0.115.0
uvicorn>=0.34.0

# HTTP Client
httpx>=0.28.0

# Google Generative AI
google-generativeai>=0.8.0

# Data validation
pydantic>=2.10.0

# Utilities
python-dotenv>=1.0.0
